Refactoring a little
